+++ kukumo-tools/jext/src/main/java/iti/commons/jext/Extension.java
@@ -87,4 +87,5 @@ public @interface Extension {
      */
     String overrides() default "";
 
+
 }