+++ kukumo/kukumo-server/src/main/java/iti/kukumo/server/ExecutionManager.java
@@ -5,41 +5,56 @@ import iti.kukumo.api.Kukumo;
 import iti.kukumo.api.KukumoConfiguration;
 import iti.kukumo.api.plan.PlanNode;
 import iti.kukumo.api.plan.PlanNodeSnapshot;
-import iti.kukumo.server.app.PlanNodeWrapper;
-import org.apache.commons.io.IOUtils;
+import iti.kukumo.server.spi.ExecutionRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import javax.annotation.PostConstruct;
 import java.io.IOException;
 import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentSkipListSet;
+import java.util.stream.Collectors;
 
 @Component
 public class ExecutionManager {
 
+    private static Map<String, PlanNode> aliveExecutions = new ConcurrentHashMap<>();
+    private static Set<String> runningExecutions = new ConcurrentSkipListSet<>();
+    private static Set<String> finishedExecutions = new ConcurrentSkipListSet<>();
 
-    private static Map<String, PlanNode> executions = new HashMap<>();
-    private static Set<String> runningExecutions = new HashSet<>();
-    private static Set<String> finishedExecutions = new HashSet<>();
+    @Value("${kukumo.executions.path}")
+    private String executionsPath;
 
-    private String newID() {
-        return UUID.randomUUID().toString();
+    @Value("${kukumo.executions.oldestAge}")
+    private int executionsOldestAge;
+
+    @Autowired
+    private ExecutionRepository executionRepository;
+
+    @PostConstruct
+    private void fillFinishedExecutions() {
+        executionRepository.removeOldExecutions(executionsOldestAge);
+        finishedExecutions.addAll(executionRepository.getAllExecutionIDs());
     }
 
 
-    public PlanNodeWrapper run(String contentType, InputStream content) {
-        String executionID = newID();
+    public KukumoExecution run(String contentType, InputStream content) {
+        String executionID = UUID.randomUUID().toString();
         runningExecutions.add(executionID);
         try {
-            Configuration configuration = KukumoConfiguration.defaultConfiguration().appendFromMap(Map.of(
+            Configuration configuration = Kukumo.defaultConfiguration().appendFromMap(Map.of(
                     KukumoConfiguration.RESOURCE_TYPES, contentType,
-                    KukumoConfiguration.OUTPUT_FILE_PATH, "/tmp/" + executionID + "/kukumo.json"
+                    KukumoConfiguration.OUTPUT_FILE_PATH, executionsPath + executionID + "/kukumo.json",
+                    KukumoConfiguration.EXECUTION_ID, executionID
             ));
             var plan = Kukumo.instance().createPlanFromContent(configuration, content);
-            executions.put(executionID,plan);
+            aliveExecutions.put(executionID,plan);
             var result = Kukumo.instance().executePlan(plan, configuration);
-            return new PlanNodeWrapper(Map.of("executionId",executionID),new PlanNodeSnapshot(plan));
+            return new KukumoExecution(configuration.asMap(),new PlanNodeSnapshot(plan));
         } finally {
             runningExecutions.remove(executionID);
             finishedExecutions.add(executionID);
@@ -47,14 +62,28 @@ public class ExecutionManager {
     }
 
 
-    public PlanNodeWrapper getExecutionData(String executionID) throws IOException {
+    public KukumoExecution getExecution(String executionID) throws IOException {
         PlanNodeSnapshot plan;
         if (runningExecutions.contains(executionID)) {
-            plan = new PlanNodeSnapshot(executions.get(executionID));
+            plan = new PlanNodeSnapshot(aliveExecutions.get(executionID));
         } else {
-            plan = Kukumo.planSerializer().read(Path.of("/tmp/"+executionID+"/kukumo.json"));
+            plan = Kukumo.planSerializer().read(Path.of(executionsPath + executionID+"/kukumo.json"));
         }
-        return new PlanNodeWrapper(Map.of("executionId",executionID),plan);
+        return new KukumoExecution(Map.of("executionId",executionID),plan);
     }
+
+
+    public List<KukumoExecution> searchExecutionHistory(ExecutionCriteria criteria) throws IOException {
+        return executionRepository.getExecutions(criteria);
+    }
+
+
+    public List<KukumoExecution> getAliveExecutions() {
+        return aliveExecutions.values().stream()
+            .map(PlanNodeSnapshot::new)
+            .map(KukumoExecution::new)
+            .collect(Collectors.toList());
+    }
+
 
 }