+++ kukumo/kukumo-server/src/test/java/iti/kukumo/server/KukumoServerApplicationTests.java
@@ -2,8 +2,10 @@ package iti.kukumo.server;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
+@ActiveProfiles("test")
 class KukumoServerApplicationTests {
 
 	@Test