+++ kukumo-tools/configurer/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog][1],
 and this project adheres to [Semantic Versioning][2].
 
 ## [Unreleased]
+
+## [1.0.1-SERVER]
+
+### Modified
+- The method `Configuration#asMap()` uses `LinkedHashMap` instead of `HashMap` to preserve 
+the original order of the properties   
   
 ## [1.0.0] 2019-04-03